porid

+++ srcs/ft_strncat.c
@@ -6,16 +6,16 @@
 /*   By: jumourot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/11 19:10:53 by jumourot          #+#    #+#             */
-/*   Updated: 2019/05/11 19:13:08 by jumourot         ###   ########.fr       */
+/*   Updated: 2019/05/13 12:14:29 by jumourot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char        *ft_strncat(char *dest, char const *src, size_t n)
+char	*ft_strncat(char *dest, char const *src, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
